Add a firmware downloader
To make keeping firmware up-to-date even easier, add a script that
downloads the latest firmware. Data about this firmware is fetched from
the github repo of this project. It is also used by
Ilo.update_rib_firmware(): if the filename passed to this function is
"latest", it will download and use the latest firmware.