Simplify the use of snmp_address_1_trapcommunity through a string
Currently snmp_address_1_trapcommunity does not actively use the version string
as of Ilo4 version 2.10

Retrieving the parameter through RIBCL xml provides an empty version string.
Sample:
<SNMP_ADDRESS_1_TRAPCOMMUNITY VERSION="" VALUE="MY_COMMUNITY"/>

So better use a string to set the parameter.
It ease the use of the parameter though command line cli (I do not even
have a clue on how to pass a dictionnary on the cli).
This patch still maintains the dictionnary compatibility.