uEFI support in get_persistent_boot
With HP Generation 9 hardware we introduce
uEFI. The available boot selection is no longer
static and will change based on each machine.
Because of this change we must know which boot selection
items are available and what each item actually is. Here
is an example of one boot selection item:
<DEVICE value="Boot0009" DESCRIPTION="Embedded FlexibleLOM 1 Port 1 : HP
       Ethernet 1Gb 4-port 331FLR Adapter - NIC (IPv4) "/>
get_persistent_boot will now return a list of tuples for
uEFI based machines.