Switch to a dual GPL/APL licensing model
At the request of HP, we switch to a dual GPL/APL model to make it
possible to use python-hpilo with OpenStack projects.

All people who contributed have indicated their approval on
https://github.com/seveas/python-hpilo/issues/85

This commit will come on top of release 2.13.1 and form a
github-tag-only 2.13.2 release as a stopgap measure, so an APL licensed
version can be released without waiting for 3.0, which is still under
development but will be delayed as I'm about to become father and will
lack time to finish it in the coming months.