Merge delayed control/info tags where possible
When doing a delayed call of e.g. factory_defaults and activate license,
we would send the following:

<RIB_INFO MODE="write">
    <FACTORY_DEFAULTS />
</RIB_INFO>
<RIB_INFO MODE="write">
    <LICENSE>
        <ACTIVATE KEY="123456789ABCDEFGHIJKLMNOP" />
    </LICENSE>
</RIB_INFO>

This can be problematic, because the iLO will start reseting itself at
the first </RIB_INFO> making the second one unreliable. So if we detect
this, let's merge the two blocks into

<RIB_INFO MODE="write">
    <FACTORY_DEFAULTS />
    <LICENSE>
        <ACTIVATE KEY="123456789ABCDEFGHIJKLMNOP" />
    </LICENSE>
</RIB_INFO>

by simply not creating a new tag in _root_element if the last tag in the
list is of the same type and with the same attributes, reusing that last
element instead.