Deal with iLO bugs that cause non-ascii data in XML
The data returned by the iLO must be ascii-only, with non-ascii
characters replaced with xml charrefs. Unfortunately, this is not done
consistently. FOr example, the serial number is not escaped, but put in
raw, so if the serialnumber is broken (I've had reports of it being set
to U+0090), the XML is broken. Work around this by detecting non-ascii
data, replacing that data and warning if we do so.